+++ SFG/spectrum/__init__.py
@@ -1,0 +1,2 @@
+from .spectrum import *
+from .averagers import *
